arreglos para flags

+++ Escuela/libft/ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: rarias-p <rarias-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/10 18:10:32 by rarias-p          #+#    #+#             */
-/*   Updated: 2019/11/15 09:24:31 by rarias-p         ###   ########.fr       */
+/*   Updated: 2019/11/17 13:49:51 by rarias-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_calloc(size_t count, size_t size)
 {
-	char	*ward;
-	int		i;
+	char			*ward;
+	unsigned long	i;
 
 	if (!(ward = malloc(count * size)))
 		return (0);
