+++ Escuela/libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: rarias-p <rarias-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 10:37:11 by rarias-p          #+#    #+#             */
-/*   Updated: 2019/11/17 13:17:39 by rarias-p         ###   ########.fr       */
+/*   Updated: 2019/11/17 13:54:51 by rarias-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*aux;
 
 	aux = (char *)s;
