+++ Escuela/libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: rarias-p <rarias-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/10 10:38:32 by rarias-p          #+#    #+#             */
-/*   Updated: 2019/11/15 09:23:38 by rarias-p         ###   ########.fr       */
+/*   Updated: 2019/11/17 13:57:33 by rarias-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	while (dest[i] != '\0')
