+++ Escuela/libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: rarias-p <rarias-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/07 11:22:56 by rarias-p          #+#    #+#             */
-/*   Updated: 2019/11/15 09:23:34 by rarias-p         ###   ########.fr       */
+/*   Updated: 2019/11/17 13:58:10 by rarias-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < size - 1 && src[i] != '\0')
