+++ Escuela/libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: rarias-p <rarias-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/10 11:19:58 by rarias-p          #+#    #+#             */
-/*   Updated: 2019/11/15 09:23:21 by rarias-p         ###   ########.fr       */
+/*   Updated: 2019/11/17 13:59:16 by rarias-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	if (needle[0] == '\0')
